+++ CHANGELOG.md
@@ -5,6 +5,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+- SqlServerDsc
+  - Added support for the Service Broker Endpoint in the SqlEndPoint
+
 ### Changed
 
 - SqlServerDsc